Add Dockerfile, run on ext. interface
Add a Dockerfile to build Docker images from the current slides.

By default, present runs the slideshow on the localhost interface.
Even if we expose the port 3999, we cannot reach present from outside
of the docker container. One way to run present on the external
interface is to grab the first IP address from /etc/hosts and run on
that interface.